xmpp: Preserve pushkitDeviceId when re-enabling
This is called in some circumstances, such as when modifying MUC push
settings. As pushkitDeviceId was not extracted and passed back to enable()
it defaulted to nil and was lost.

This triggered a push re-register on subsequent connections (because the
stored pushkitDeviceId (nil) did not match the one from the system).