MessageEventHandler: don't skip notifications for archived messages
This is an attempt to fix https://github.com/snikket-im/snikket-ios/issues/161

There are very likely to be cases where this will notify about messages that
should not be notified about, but that seems better right now than *not*
notifying about messages we don't receive a push notification for.

Push notifications may not arrive if the device is offline, and then opening
the app later should still show the received messages as unread.

Hopefully we can move to a more robust method for determining which messages
and chats are unread, but for now this is a simple workaround worth trying.